backend/requirements.txt
Django>=5.1,<6.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
 psycopg[binary]>=3.2
 celery[redis]>=5.5
 redis>=5.0
django-filter>=24.2
drf-spectacular>=0.27
django-storages[boto3]>=1.14
 boto3>=1.35
 channels>=4.1
 channels-redis>=4.2
 stripe>=11.0
 sentry-sdk>=2.14
 python-dotenv>=1.0
 gunicorn>=23.0
 pytest>=8.3
 pytest-django>=4.9
 factory-boy>=3.3
